Make test_deepseek_v3.py a benchmark as well as a test
