+++ python/nvfuser/testing/benchmark_utils.py
@@ -1,0 +1,24 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-present NVIDIA CORPORATION & AFFILIATES.
+# All rights reserved.
+# SPDX-License-Identifier: BSD-3-Clause
+
+import torch
+
+def get_benchmark_fn(func, /, profile: bool):
+    def wrapper(*args, **kwargs):
+        if profile:
+            torch.cuda.cudart().cudaProfilerStart()
+        result = func(*args, **kwargs)
+        torch.cuda.synchronize()
+        if profile:
+            torch.cuda.cudart().cudaProfilerStop()
+        return result
+
+    return wrapper
+
+
+# Returns two functors, the first with profiler off and the second with profiler
+# on. The first functor is usually used for warmup and the second for actual
+# benchmarking.
+def get_benchmark_fns(func):
+    return get_benchmark_fn(func, profile=False), get_benchmark_fn(func, profile=True)
